+++ 09_Texturing/Intersection.h
@@ -9,16 +9,16 @@ private:
 	Material material;
 	Vec3 normal;
 	std::optional<TextureCoordinates> texCoords;
-	double t;
+	float t;
 
 public:
-	Intersection(const Material& material, const Vec3& normal, const std::optional<TextureCoordinates>& texCoords, double t)
+	Intersection(const Material& material, const Vec3& normal, const std::optional<TextureCoordinates>& texCoords, float t)
 		: material(material), normal(normal), texCoords(texCoords), t(t)
 	{ };
 
 	Material getMaterial() const;
 	Vec3 getNormal() const;
-	double getT() const;
+	float getT() const;
 	std::optional<TextureCoordinates> getTexCoords() const;
 
 };
