+++ 09_Texturing/Plane.cpp
@@ -7,11 +7,11 @@ Material Plane::getMaterial() const
 
 std::optional<Intersection> Plane::intersect(const Ray& ray) const
 {
-    double denom = Vec3::dot(ray.getDirection(), normal);
+    float denom = Vec3::dot(ray.getDirection(), normal);
     if (denom == 0.0)
         return {};
 
-    double t = -(Vec3::dot(ray.getOrigin(), normal) + d) / denom;
+    float t = -(Vec3::dot(ray.getOrigin(), normal) + d) / denom;
     if (t < 0)
         return {};
